+++ lab_6/functions.py
